+++ packages/yourloops/test/integration/pages/patient-view/patient-profile/patient-profile-view.hcp.spec.tsx
@@ -34,7 +34,7 @@ import { mockPatientApiForHcp } from '../../../mock/patient.api.mock'
 import { mockDataAPI } from '../../../mock/data.api.mock'
 import { AppMainLayoutHcpParams, testAppMainLayoutForHcp } from '../../../use-cases/app-main-layout-visualisation'
 import { renderPage } from '../../../utils/render'
-import { patient1Id, patientWithMmolId } from '../../../data/patient.api.data'
+import { patient1Id, patient2Id, patient3Id, patientWithMmolId } from '../../../data/patient.api.data'
 import { AppUserRoute } from '../../../../../models/enums/routes.enum'
 import {
   testMonitoringAlertsParametersConfigurationForPatientMgdl,
@@ -57,6 +57,13 @@ import userEvent from '@testing-library/user-event'
 import { getTranslation } from '../../../../utils/i18n'
 import { mockDblCommunicationApi } from '../../../mock/dbl-communication.api'
 import { checkProfilesForMalePatient } from '../../../assert/profile-range.assert'
+import {
+  checkCliniciansEmptyList, checkCliniciansFiveClinicians,
+  checkCliniciansManagementErrors, checkCliniciansManagementHcp
+} from '../../../use-cases/clinicians-management'
+import { mockLeadCliniciansApi } from '../../../mock/clinicians.api.mock'
+import { LeadCliniciansApi } from '../../../../../lib/lead-clinicians/lead-clinicians.api'
+import ErrorApi from '../../../../../lib/error/error.api'
 
 describe('Patient profile view for HCP', () => {
   beforeEach(() => {
@@ -68,6 +75,7 @@ describe('Patient profile view for HCP', () => {
     mockPatientApiForHcp()
     mockDataAPI()
     mockDblCommunicationApi()
+    mockLeadCliniciansApi()
   })
 
   const firstName = 'HCP firstName'
@@ -76,6 +84,9 @@ describe('Patient profile view for HCP', () => {
   const patientTargetAndAlertsRouteMmoL = `/teams/${myThirdTeamId}/patients/${patientWithMmolId}${AppUserRoute.PatientProfile}`
   const patientProfileRoute = `/teams/${myThirdTeamId}/patients/${patient1Id}${AppUserRoute.PatientProfile}`
   const malePatientProfileRoute = `/teams/${myThirdTeamId}/patients/${patientWithMmolId}${AppUserRoute.PatientProfile}`
+
+  const patientWithOneClinicianRoute = `/teams/${myThirdTeamId}/patients/${patient2Id}${AppUserRoute.PatientProfile}`
+  const patientWithFiveCliniciansRoute = `/teams/${myThirdTeamId}/patients/${patient3Id}${AppUserRoute.PatientProfile}`
 
   /**
    * @see https://github.com/testing-library/react-testing-library/issues/651
@@ -162,6 +173,42 @@ describe('Patient profile view for HCP', () => {
 
       const ageText = await screen.findByText(/age/i)
       expect(ageText).toBeInTheDocument()
+    })
+
+    it('should be able to view clinicians list with 0 clinician', async () => {
+      await act(async () => {
+        renderPage(patientProfileRoute)
+      })
+
+      await checkCliniciansEmptyList()
+    })
+
+    it('should be able to view and manage clinicians list with 1 clinician', async () => {
+      await act(async () => {
+        renderPage(patientWithOneClinicianRoute)
+      })
+
+      await checkCliniciansManagementHcp()
+    })
+
+    it('should be able to view clinicians list with 5 clinicians', async () => {
+      await act(async () => {
+        renderPage(patientWithFiveCliniciansRoute)
+      })
+
+      await checkCliniciansFiveClinicians()
+    })
+
+    it('should handle gracefully error cases when managing clinicians', async () => {
+      jest.spyOn(LeadCliniciansApi, 'addClinician').mockRejectedValue('Add clinician error')
+      jest.spyOn(LeadCliniciansApi, 'removeClinician').mockRejectedValue('Remove clinician error')
+      jest.spyOn(ErrorApi, 'sendError').mockResolvedValue()
+
+      await act(async () => {
+        renderPage(patientWithOneClinicianRoute)
+      })
+
+      await checkCliniciansManagementErrors()
     })
 
     // TODO: Uncomment when the feature is implemented with the API