+++ CHANGELOG.md
@@ -8,6 +8,8 @@ It is based on Tidepool Blip 1.27.
 - YLP-3614 Hcps can acknowledge patient monitoring alerts
 - YLP-3617 As a user, I can see the devices changes history
 - YLP-3785 Be able to know the effective DBLG1 or DBLG2 settings for the displayed day in the daily view
+- YLP-3787 Referring HCP management for patients
+- YLP-3788 Referring HCP management for HCPs
 - YLP-3813 Add hyperglycemia alerts (display in the patient view and settings in care teams and patients profile)
 - YLP-3837 Display Dexcom G7 sensor information in device view